+++ docs/changelog.rst
@@ -12,6 +12,10 @@ Adjusted to changes in the clkhash library.
 
 Implemented in #684
 
+**Delete encodings from database at project deletion**
+
+Encodings will be deleted at project deletion, but only for projects created with this version or higher.
+
 Version 1.15.0
 ---------------