+++ docs/changelog.rst
@@ -6,6 +6,12 @@ Changelog
 
 Next Version
 ------------
+**Delete upload files on object store after ingestion**
+
+If a data provider uploads its data via the object store, we now clean up afterwards.
+
+Implemented in #686
+
 **fixed Record Linkage API tutorial**
 
 Adjusted to changes in the clkhash library.
@@ -15,6 +21,8 @@ Implemented in #684
 **Delete encodings from database at project deletion**
 
 Encodings will be deleted at project deletion, but only for projects created with this version or higher.
+
+Implemented in #683
 
 Version 1.15.0
 ---------------