+++ docs/changelog.rst
@@ -33,6 +33,7 @@ Encodings will be deleted at project deletion, but only for projects created wit
 
 Implemented in #683
 
+
 Version 1.15.0
 ---------------
 
@@ -65,6 +66,7 @@ The run status endpoint now includes `total_number_of_comparisons` for completed
 Implemented in: #651
 
 As usual lots of version upgrades - now using the latest stable redis and postgresql.
+
 
 Version 1.14.0
 ---------------